+++ Common/Advanced/EarlyAccess/IRoslynCompilationGuard.cs
@@ -1,7 +1,8 @@
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Text;
 
 namespace MirrorSharp.Advanced.EarlyAccess {
-    internal interface IRoslynGuard {
+    internal interface IRoslynCompilationGuard {
         void ValidateCompilation(Compilation compilation, IRoslynSession session);
     }
 }