+++ Common/Advanced/EarlyAccess/IRoslynSourceTextGuard.cs
@@ -1,0 +1,7 @@
+using Microsoft.CodeAnalysis.Text;
+
+namespace MirrorSharp.Advanced.EarlyAccess {
+    internal interface IRoslynSourceTextGuard {
+        void ValidateSourceText(SourceText sourceText);
+    }
+}