+++ Owin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [2.3.3] - 2020-03-04
+
+### Changed
+- Internal changes (exposed some internals to SharpLab)
+
+## [2.3.2] - 2020-03-04
+
+### Changed
+- Updated to support MirrorSharp.Common 2.2.3
+
 ## [2.3.1] - 2020-12-17
 
 ### Changed