+++ VisualBasic/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.1.2] - 2020-12-17
+
+### Changed
+- Updated to support MirrorSharp.Common 2.2.3
+
 ## [2.1.1] - 2020-12-17
 
 ### Changed